Move EventDispatcher to their own sub-namespace
