+++ src/AlternateLife.RageMP.Net/Scripting/ScriptingClasses/EventScripting.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using AlternateLife.RageMP.Net.Enums;
 using AlternateLife.RageMP.Net.EventArgs;
 using AlternateLife.RageMP.Net.Helpers;
+using AlternateLife.RageMP.Net.Helpers.EventDispatcher;
 using AlternateLife.RageMP.Net.Interfaces;
 using AlternateLife.RageMP.Net.Native;
